Serialize timestamp as a JSON Number, not String
In the GELF payload specification, this is specified to be a JSON number.
Graylog (tested on v2.5.2) will accept JSON string values, but emit a warning in
the server logs that the message has an invalid timestamp. Avoid this by
converting the timestamp to f64 then serializing it. This does introduce the
possibility of floating-point inaccuracy, but:
- Graylog converts incoming timestamps to doubles anyway
- serde_json by default uses f64 internally for numbers
- f64 can represent timestamps accurate to a second or less until the year 2255